requirements.txt
robotframework
robotframework-seleniumlibrary
robotframework-browser
selenium
pandas
openpyxl

google-auth
gspread

robotframework-metrics